myeskwela-aws-master/myeskwela/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
bidict==0.21.4
blinker==1.4
cbor2==5.4.2.post1
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
cloudant==2.15.0
cloudinary==1.29.0
dnspython==2.2.1
et-xmlfile==1.1.0
eventlet==0.33.0
Flask==2.0.3
Flask-HTTPAuth==4.5.0
Flask-JSGlue==0.3.1
Flask-Mail==0.9.1
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
frozenlist==1.3.0
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.0
jdcal==1.4.1
Jinja2==3.0.3
MarkupSafe==2.1.0
monotonic==1.6
multidict==6.0.2
openpyxl==3.0.9
Pillow==9.0.1
psycopg2-binary==2.9.3
pubnub==6.1.0
pycryptodomex>=3.14.1
python-engineio==4.3.1
python-socketio==5.5.2
qrcode==7.3.1
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.32
urllib3==1.26.8
validate-email==1.3
Werkzeug==2.0.3
yarl==1.7.2
zope.event==4.5.0
zope.interface==5.4.0
